home/streamer/requirements.txt
# Tom's Streamer v3.0 — Python dependencies
# Instalacja: pip3 install -r requirements.txt --break-system-packages

# Web framework
flask>=3.0
flask-socketio>=5.3
eventlet>=0.35          # lub gevent (zmień async_mode w app.py)

# GStreamer Python bindings
# Instalacja przez apt, nie pip:
# sudo apt install python3-gi python3-gi-cairo gir1.2-gstreamer-1.0
# sudo apt install gstreamer1.0-plugins-good gstreamer1.0-plugins-bad
# sudo apt install gstreamer1.0-plugins-ugly gstreamer1.0-alsa
# sudo apt install gstreamer1.0-libav

# Bluetooth
dbus-python>=1.3        # apt: python3-dbus

# System monitoring
psutil>=5.9

# Serial (UART RP2040)
pyserial>=3.5

Pillow
requests